% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileToMatrix.R
\name{fileToMatrix}
\alias{fileToMatrix}
\title{function to get a ine file and returns a numerical matrix A.}
\usage{
fileToMatrix(path)
}
\arguments{
\item{path}{A string that containes the path to an ine or a ext file. The ine file desrcibes a H-polytope and ext file describes a V-polytope or a zonotope.}
}
\value{
If the path corresponds to an ine file then the return value is a list that containes elements "A" and "b", i.e. the numerical \eqn{m\times d} matrix \eqn{A} and the numerical \eqn{m}-dimensional vector \eqn{b}, defining H-polytope \eqn{P}, s.t.:  \eqn{Ax\leq b}. If it corresponds to an ext file (V-polytopes or zonotopes) then the return value is a \eqn{m\times d} matrix that containes row-wise the vertices or the segments respectively.
}
\description{
This function takes the path for an ine or an ext file and returns the corresponding numerical matrix and vector that are compatible with volesti package's functions.
}
\examples{
# give the path to birk4.ine
ListPoly = fileToMatrix(path/to/data/birk4.ine)
}
