% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSkinnyCube.R
\name{GenSkinnyCube}
\alias{GenSkinnyCube}
\title{Generator function for skinny hypercubes.}
\usage{
GenSkinnyCube(dimension, repr = "H")
}
\arguments{
\item{dimension}{The dimension of the skinny hypercube.}
}
\value{
A \eqn{d}-dimensional skinny hypercube in H-representation. The retutn value is a list with two elements: the "matrix" containing a \eqn{2d\times d} matrix \eqn{A} and the "vector" containing a \eqn{2d}-dimensional vector \eqn{b}, s.t. \eqn{Ax\leq b}.
}
\description{
This function can be used to generate a \eqn{d}-dimensional skinny hypercube only in H-representation.
}
\examples{
# generate a 10-dimensional skinny hypercube.
PolyList = GenSkinnyCube(10)
}
