% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineToMatrix.R
\name{ineToMatrix}
\alias{ineToMatrix}
\title{function to get a ine file and returns a numerical matrix A.}
\usage{
ineToMatrix(P)
}
\arguments{
\item{P}{It is in format, read.cs('path/to/file.ine'). The ine file desrcibes a H-polytope.}
}
\value{
The numerical matrix in ine format.
}
\description{
This function takes an ine file as a string (using read.csv()) and returns a numerical matrix \eqn{A} in ine format for function volume (see \eqn{volume} function examples).
}
\examples{
# give the path to birk4.ine
A = ineToMatrix(read.csv('path/to/data/birk4.ine'))
}
