% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeSdpaFormatFile}
\alias{writeSdpaFormatFile}
\title{Write a SDPA format file}
\usage{
writeSdpaFormatFile(
  spectrahedron = NULL,
  objectiveFunction = NULL,
  outputFile = NULL
)
}
\arguments{
\item{spectrahedron}{A spectrahedron in n dimensions; must be an object of class Spectrahedron}

\item{objectiveFunction}{A numerical vector of length n}

\item{outputFile}{Name of the output file}
}
\description{
Outputs a spectrahedron (the matrices defining a linear matrix inequality) and a vector (the objective function)
to a SDPA format file.
}
