% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZdemoVolume.R
\name{ZdemoVolume}
\alias{ZdemoVolume}
\title{Run some volume approximation experiments for zonotopes.}
\usage{
ZdemoVolume(algo)
}
\arguments{
\item{CG}{The string "CG" to choose CoolingGaussian algorithm.}

\item{SOB}{The string "SOB" to choose SequenceOfBalls algorithm.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Run SequenceOfBalls or CoolingGaussian algorithm to approximate the volume of some zonotopes. In each test we use GenZonotope() function to generate a random zonotope and then we apply rounding before the volume approximation.
For each polytope we run \eqn{10} volume experiments and we consider the mean value as the volume approximation. For SOB algorithm we demand \eqn{error = 0.1} and for CG algorithm we demand \eqn{error = 0.2}.
}
\examples{
# test SequenceOfBalls
ZdemoVolume("SOB")
# test CoolingGausian
ZdemoVolume("CG")
}
